+++ pvinspect/integration/pytorch/__init__.py
@@ -1,1 +1,1 @@
-from .dataset import Dataset
+from .dataset import Dataset, ClassificationDataset