+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Pandas methods are now (all) available from `ImageSequence.pandas`
 * `Image` now supports lazy loading of data (which is now default)
 * Introduced convenience method `from_self` to avoid using `type(self).from_other(self, ...)`
-* Add PyTorch `Dataset` implementation
+* Add PyTorch `Dataset` and `ClassificationDataset` implementation
 * Speed up meta data handling
 * Add ELPV cell image dataset